+++ src/dashboard/apigateway/apigateway/biz/resource.py
@@ -16,8 +16,6 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-
-
 import json
 from typing import List, Optional
